front: split too long files of infra editor
