+++ front/src/applications/editor/tools/pointEdition/consts.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const POINT_LAYER_ID = 'pointEditionTool/new-entity';
