+++ front/src/applications/editor/tools/selection/components/SelectionLeftPanel.tsx
@@ -1,110 +1,16 @@
 import React, { useContext, useEffect } from 'react';
-import { Layer, Popup, Source } from 'react-map-gl/maplibre';
-import { useSelector } from 'react-redux';
 import { groupBy, map } from 'lodash';
 import { GoNoEntry } from 'react-icons/go';
 import { RiFocus3Line } from 'react-icons/ri';
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router-dom';
 
-import type { Zone } from 'types';
+import EditorContext from '../../../context';
+import EntitySumUp from '../../../components/EntitySumUp';
+import type { SelectionState } from '../types';
+import type { ExtendedEditorContextType } from '../../editorContextTypes';
 
-import { zoneToFeature } from 'utils/mapHelper';
-
-import EditorContext from 'applications/editor/context';
-import 'applications/editor/tools/selection/styles.scss';
-import EntitySumUp from 'applications/editor/components/EntitySumUp';
-import type { SelectionState } from 'applications/editor/tools/selection/types';
-import type {
-  EditorContextType,
-  ExtendedEditorContextType,
-} from 'applications/editor/tools/editorContextTypes';
-
-import colors from 'common/Map/Consts/colors';
-import GeoJSONs from 'common/Map/Layers/GeoJSONs';
-import { useInfraID } from 'common/osrdContext';
-
-import { getMap } from 'reducers/map/selectors';
-
-export const SelectionMessages = () => {
-  const { t, state } = useContext(EditorContext) as EditorContextType<SelectionState>;
-
-  return t(`Editor.tools.select-items.help.${state.selectionState.type}-selection`).toString();
-};
-
-const SelectionZone = ({ newZone }: { newZone: Zone }) => (
-  <Source type="geojson" data={zoneToFeature(newZone)} key="new-zone">
-    <Layer type="line" paint={{ 'line-color': '#666', 'line-dasharray': [3, 3] }} />
-  </Source>
-);
-
-export const SelectionLayers = () => {
-  const {
-    state,
-    editorState: { editorLayers },
-    renderingFingerprint,
-  } = useContext(EditorContext) as ExtendedEditorContextType<SelectionState>;
-  const { mapStyle, layersSettings, issuesSettings } = useSelector(getMap);
-
-  const infraID = useInfraID();
-
-  let selectionZone: Zone | undefined;
-
-  if (state.mousePosition) {
-    if (state.selectionState.type === 'rectangle' && state.selectionState.rectangleTopLeft) {
-      selectionZone = {
-        type: 'rectangle',
-        points: [state.selectionState.rectangleTopLeft, state.mousePosition],
-      };
-    } else if (
-      state.selectionState.type === 'polygon' &&
-      state.selectionState.polygonPoints.length
-    ) {
-      selectionZone = {
-        type: 'polygon',
-        points: state.selectionState.polygonPoints.concat([state.mousePosition]),
-      };
-    }
-  }
-
-  return (
-    <>
-      <GeoJSONs
-        colors={colors[mapStyle]}
-        selection={state.selection.map((e) => e.properties.id)}
-        layers={editorLayers}
-        fingerprint={renderingFingerprint}
-        layersSettings={layersSettings}
-        issuesSettings={issuesSettings}
-        infraID={infraID}
-      />
-      {selectionZone && <SelectionZone newZone={selectionZone} />}
-      {state.mousePosition && state.selectionState.type === 'single' && state.hovered && (
-        <Popup
-          className="popup editor-selection"
-          anchor="bottom"
-          longitude={state.mousePosition[0]}
-          latitude={state.mousePosition[1]}
-          closeButton={false}
-        >
-          <EntitySumUp
-            key={state.hovered.id}
-            id={state.hovered.id}
-            objType={state.hovered.type}
-            error={state.hovered.error}
-            status={
-              state.selection.find(
-                (item) => item.properties.id === (state.hovered?.id as string)
-              ) && 'selected'
-            }
-          />
-        </Popup>
-      )}
-    </>
-  );
-};
-
-export const SelectionLeftPanel = () => {
+const SelectionLeftPanel = () => {
   const { t } = useTranslation();
   const { state, setState } = useContext(
     EditorContext
@@ -182,7 +88,7 @@ export const SelectionLeftPanel = () => {
   return (
     <>
       <h4>{t('Editor.tools.select-items.title')}</h4>
-      <ul className="list-unstyled">
+      <ul className="list-unstyled selection-left-panel">
         {selection.map((item) => (
           <li key={item.properties.id} className="pb-4">
             <div className="pb-2 entity">
@@ -217,3 +123,5 @@ export const SelectionLeftPanel = () => {
     </>
   );
 };
+
+export default SelectionLeftPanel;
