+++ front/src/applications/editor/tools/selection/components/index.ts
@@ -1,0 +1,3 @@
+export { default as SelectionLayers } from './SelectionLayers';
+export { default as SelectionLeftPanel } from './SelectionLeftPanel';
+export { default as SelectionMessages } from './SelectionMessages';
