--- front/src/applications/editor/tools/selection/styles.scss
@@ -1,12 +1,0 @@
-.entity {
-  .strong {
-    font-weight: bold;
-  }
-}
-
-.popup.editor-selection {
-  .maplibregl-popup-content {
-    min-width: 200px;
-    min-height: 100px;
-  }
-}
