+++ front/src/applications/editor/tools/switchEdition/components/index.ts
@@ -1,0 +1,3 @@
+export { default as SwitchEditionLayers } from './SwitchEditionLayers';
+export { default as SwitchEditionLeftPanel } from './SwitchEditionLeftPanel';
+export { default as SwitchMessages } from './SwitchMessages';
