+++ front/src/applications/editor/tools/trackEdition/components/index.ts
@@ -1,0 +1,3 @@
+export { default as TrackEditionLayers } from './TrackEditionLayers';
+export { default as TrackEditionLeftPanel } from './TrackEditionLeftPanel';
+export { default as TrackEditionMessages } from './TrackEditionMessages';
