+++ front/src/applications/editor/tools/trackEdition/consts.ts
@@ -1,0 +1,2 @@
+export const TRACK_LAYER_ID = 'trackEditionTool/new-track-path';
+export const POINTS_LAYER_ID = 'trackEditionTool/new-track-points';
